requirements.txt
requests[security]==2.8.1
flask==0.10.1
flask-restful==0.3.5
python-keystoneclient==2.0.0
python-novaclient==3.1.0
